refactor: 디렉토리 정리, globalSafeArea 추가
