--- lib/constants/colors.dart
@@ -1,32 +1,0 @@
-import 'package:flutter/material.dart';
-
-class BaseColors {
-  static const green = {
-    0: Color(0xff2EE48D),
-    1: Color(0xff6AD975),
-    2: Color(0xff28BAB1),
-  };
-
-  static const blue = {
-    0: Color(0xff1DCBDB),
-    1: Color(0xff0D57C4),
-  };
-
-  static const red = {
-    0: Color(0xffED0000),
-  };
-
-  static const black = {
-    0: Color(0xffF8F8F8),
-    1: Color(0xffF4F4F4),
-    2: Color(0xffE5E5E5),
-    3: Color(0xffD5D5D5),
-    4: Color(0xffCACACA),
-    5: Color(0xffC1C1C1),
-    6: Color(0xffABABAB),
-    7: Color(0xff878787),
-    8: Color(0xff616161),
-    9: Color(0xff3E3E3E),
-    10: Color(0xff181818),
-  };
-}
