--- lib/constants/sizes.dart
@@ -1,32 +1,0 @@
-class Sizes {
-  static const size1 = 1.0;
-  static const size2 = 2.0;
-  static const size3 = 3.0;
-  static const size4 = 4.0;
-  static const size5 = 5.0;
-  static const size6 = 6.0;
-  static const size7 = 7.0;
-  static const size8 = 8.0;
-  static const size9 = 9.0;
-  static const size10 = 10.0;
-  static const size11 = 11.0;
-  static const size12 = 12.0;
-  static const size14 = 14.0;
-  static const size16 = 16.0;
-  static const size18 = 18.0;
-  static const size20 = 20.0;
-  static const size24 = 24.0;
-  static const size28 = 28.0;
-  static const size32 = 32.0;
-  static const size36 = 36.0;
-  static const size40 = 40.0;
-  static const size44 = 44.0;
-  static const size48 = 48.0;
-  static const size52 = 52.0;
-  static const size56 = 56.0;
-  static const size60 = 60.0;
-  static const size64 = 64.0;
-  static const size72 = 72.0;
-  static const size80 = 80.0;
-  static const size96 = 96.0;
-}
