+++ lib/constants/styles.dart
@@ -1,5 +1,68 @@
 import 'package:flutter/material.dart';
-import 'package:mutyne/constants/sizes.dart';
+
+class BaseColors {
+  static const green = {
+    0: Color(0xff2EE48D),
+    1: Color(0xff6AD975),
+    2: Color(0xff28BAB1),
+  };
+
+  static const blue = {
+    0: Color(0xff1DCBDB),
+    1: Color(0xff0D57C4),
+  };
+
+  static const red = {
+    0: Color(0xffED0000),
+  };
+
+  static const black = {
+    0: Color(0xffF8F8F8),
+    1: Color(0xffF4F4F4),
+    2: Color(0xffE5E5E5),
+    3: Color(0xffD5D5D5),
+    4: Color(0xffCACACA),
+    5: Color(0xffC1C1C1),
+    6: Color(0xffABABAB),
+    7: Color(0xff878787),
+    8: Color(0xff616161),
+    9: Color(0xff3E3E3E),
+    10: Color(0xff181818),
+  };
+}
+
+class Sizes {
+  static const size1 = 1.0;
+  static const size2 = 2.0;
+  static const size3 = 3.0;
+  static const size4 = 4.0;
+  static const size5 = 5.0;
+  static const size6 = 6.0;
+  static const size7 = 7.0;
+  static const size8 = 8.0;
+  static const size9 = 9.0;
+  static const size10 = 10.0;
+  static const size11 = 11.0;
+  static const size12 = 12.0;
+  static const size14 = 14.0;
+  static const size16 = 16.0;
+  static const size18 = 18.0;
+  static const size20 = 20.0;
+  static const size24 = 24.0;
+  static const size28 = 28.0;
+  static const size32 = 32.0;
+  static const size36 = 36.0;
+  static const size40 = 40.0;
+  static const size44 = 44.0;
+  static const size48 = 48.0;
+  static const size52 = 52.0;
+  static const size56 = 56.0;
+  static const size60 = 60.0;
+  static const size64 = 64.0;
+  static const size72 = 72.0;
+  static const size80 = 80.0;
+  static const size96 = 96.0;
+}
 
 class Gaps {
   // Vertical Gaps
