+++ lib/features/authentication/views/login_screen.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:mutyne/constants/gaps.dart';
-import 'package:mutyne/constants/sizes.dart';
-import 'package:mutyne/constants/colors.dart';
+import 'package:mutyne/constants/styles.dart';
+import 'package:mutyne/common/widgets/global_safe_area.dart';
 
 class LoginScreen extends StatefulWidget {
   static String routeName = "login";
@@ -20,17 +19,13 @@ class _LoginScreenState extends State<LoginScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        leading: GestureDetector(
-          onTap: () => _onNavigateBack(context),
-          child: Image.asset('assets/arrow_back.png'),
-        ),
-      ),
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(
-            horizontal: Sizes.size20,
+        appBar: AppBar(
+          leading: GestureDetector(
+            onTap: () => _onNavigateBack(context),
+            child: Image.asset('assets/arrow_back.png'),
           ),
+        ),
+        body: GlobalSafeArea(
           child: Column(children: [
             Gaps.v40,
             // 여기 Container안에서 Login ui 작업하신거 연결하면 되요!
@@ -41,8 +36,6 @@ class _LoginScreenState extends State<LoginScreen> {
               child: const Center(child: Text('로그인 UI 작업')),
             )
           ]),
-        ),
-      ),
-    );
+        ));
   }
 }
